+++ tests/Rule/AnyRuleTest.php
@@ -98,6 +98,13 @@ final class AnyRuleTest extends RuleTestCase
         return [
             'right away' => [1, new AnyRule([new IntegerType(), new FloatType()])],
             'later' => [1.5, new AnyRule([new IntegerType(), new FloatType()])],
+            'using as attribute' => [
+                new class () {
+                    #[AnyRule([new IntegerType(), new FloatType()])]
+                    private int|float $sum = 1.5;
+                },
+                null,
+            ],
         ];
     }
 
@@ -107,6 +114,14 @@ final class AnyRuleTest extends RuleTestCase
 
         return [
             'none' => ['1', new AnyRule([new IntegerType(), new FloatType()]), ['' => [$message]]],
+            'using as attribute' => [
+                new class () {
+                    #[AnyRule([new IntegerType(), new FloatType()])]
+                    private string $sum = '1.5';
+                },
+                null,
+                ['sum' => [$message]],
+            ],
         ];
     }
