+++ tests/Rule/CompositeTest.php
@@ -19,6 +19,8 @@ use Yiisoft\Validator\Tests\Rule\Base\SkipOnErrorTestTrait;
 use Yiisoft\Validator\Tests\Rule\Base\WhenTestTrait;
 use Yiisoft\Validator\Tests\Support\Rule\CoordinatesRuleSet;
 use Yiisoft\Validator\Tests\Support\Rule\RuleWithoutOptions;
+use Yiisoft\Validator\Tests\Support\Data\CompositeWithCallbackAttribute;
+use Yiisoft\Validator\Validator;
 
 final class CompositeTest extends RuleTestCase
 {
@@ -347,6 +349,19 @@ final class CompositeTest extends RuleTestCase
     {
         $when = static fn (mixed $value): bool => $value !== null;
         $this->testWhenInternal(new Composite([]), new Composite([], when: $when));
+    }
+
+    public function testWithCallbackAttribute(): void
+    {
+        $result = (new Validator())->validate(new CompositeWithCallbackAttribute());
+
+        $this->assertSame(
+            [
+                '' => ['Invalid A.'],
+                'b' => ['Invalid B.'],
+            ],
+            $result->getErrorMessagesIndexedByProperty()
+        );
     }
 
     protected function getDifferentRuleInHandlerItems(): array
